Move stdlib functions to own directory and modules
