+++ src/stdlib/io.js
@@ -1,0 +1,8 @@
+// I/O functions
+const print = (...args) => {
+  console.log(...args);
+};
+
+module.exports = {
+  print,
+};
