+++ src/stdlib/utils.js
@@ -1,0 +1,5 @@
+const all = (fn) => (...list) => list.reduce(fn);
+
+module.exports = {
+  all,
+};
